functions/requirements.txt
firebase_functions~=0.1.0
firebase_admin
msal
requests
python-dotenv